kernel: fix thread prio and stack size types in some APIs
Prio should be an int, since values are small integers, not a fixed-size
int32_t. It aligns with the prio parameters of the other APIs.

Stack size should be size_t.

+++ arch/arm/core/thread.c
@@ -80,7 +80,7 @@ static ALWAYS_INLINE void thread_monitor_init(struct tcs *tcs)
  * @return N/A
  */
 
-void _new_thread(char *pStackMem, unsigned stackSize,
+void _new_thread(char *pStackMem, size_t stackSize,
 		 void *uk_task_ptr, _thread_entry_t pEntry,
 		 void *parameter1, void *parameter2, void *parameter3,
 		 int priority, unsigned options)
