+++ arch/x86/core/thread.c
@@ -245,7 +245,7 @@ __asm__("\t.globl _thread_entry\n"
  *
  * @return opaque pointer to initialized k_thread structure
  */
-void _new_thread(char *pStackMem, unsigned stackSize,
+void _new_thread(char *pStackMem, size_t stackSize,
 		 void *uk_task_ptr, _thread_entry_t pEntry,
 		 void *parameter1, void *parameter2, void *parameter3,
 		 int priority, unsigned options)
